% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanflowering.R
\name{cleanflowering}
\alias{cleanflowering}
\title{Read and clean flowering time from experiment spatial grid raw data}
\usage{
cleanflowering(data = longform.rm)
}
\arguments{
\item{data}{long format flowering data}
}
\description{
Read and clean flowering time from experiment spatial grid raw data
}
\details{
A number of flags were used during flowering time accquisition. This function just considers everything that had these values as "dead" (or that never germinated). This is passed to make_data_flowering where it will be filtered
}

