% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{data}
\name{field}
\alias{field}
\title{Data from a multi-site field experiment with 517 sequenced Arabidopsis thaliana accessions}
\format{A data frame with  ~30,000 rows and ~25 variables:
\describe{
  \item{id}{the 1001 Genomes IDs of the used accessions}
  \item{name}{common name of the accession}
  \item{country}{of origin}
  \item{latitude}{in degrees}
  \item{longitude}{in degrees}
  \item{kgroup}{genetic grups from ADMIXTURE analysis}
  \item{qp}{8x5 cells quickpot tray }
  \item{pos}{position coordinate within the 8x5 cells qp tray}
  \item{site}{experiment location, Madrid or Tuebingen}
  \item{rep}{replicate block within experiment}
  \item{trayid}{the block -- depends on replicate and treatments}
  \item{indpop}{type of replicate. Individual or an experimental population with maximum 30 individuals}
  \item{water}{watering treatment, high or low}

  \item{FT.q}{quality of flowering time observation. 0=dead, 1=survived, -9=potential error}
  \item{FT.date}{date assigned by experimenter in the flowering time monitoring}
  \item{FT.dif}{flowering time in days from the day of sowing till FT.date}

  \item{pathimage}{where is the image from which the information in the following collumns was extracted}
  \item{datalabeled}{date where the image was labeled by a human visually inspecting the information in the image}
  \item{organ}{plant organ, either rosette or inflorescence}
  \item{ptype}{replicate type, should be identical to indpop column}
  \item{num}{number of individuals -- out of the 30 seeds sown -- survived till the end. When it is individual replicate NA is indicated.
  When it is population but failed to count it, -9 is indicated}
  \item{Harv.q}{quality of harvested plant. 0= no information existed, 1= was labeled and picture exists or written in the diary}
  \item{Harv.area}{total area of the plant organ using Otsu's segmentation algorithm}
  \item{Harv.sk}{total length of sekeleton of the plant organ. most informative for inflorescences}
  \item{Harv.bp}{number of branching points detected using a hit-or-miss algorithm with customized structural elements}
  \item{Harv.ep}{number of end points detected using a hit-or-miss algorithm with customized structural elements}
  \item{Harv.fru}{estimated number of fruits from building a multivariate regression of the above parameters with manually counted fruits}

}}
\source{
see \url{1001genomes.org/} for accession list, see \url{http://biorxiv.org/content/early/2017/03/19/118067} for genetic groups.
}
\usage{
field
}
\description{
A dataset containing different life history phenotypes and fitness estimates
of 517 accessions with 24 replicates each in different treatments, namely
site (Madrid or Tuebingen), watering (high and low), and density of plants
per pot (an experimental population or single individual).
}
\keyword{datasets}

