% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldpgk.R
\docType{package}
\name{fieldpackage}
\alias{fieldpackage}
\alias{fieldpackage-package}
\title{field: A package for}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{functions}{

...
}

\section{datasets}{

....
}
\seealso{
\url{http://r-pkgs.had.co.nz/man.html} for package documentation
}

