% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypepositions.R
\name{genotypepositions}
\alias{genotypepositions}
\title{Get the genotype positions in blocks, trays, and within-tray of the two field experiments}
\usage{
genotypepositions(location = "tuebingen", QC = TRUE)
}
\arguments{
\item{QC}{logical whether quality filter should be done or not}

\item{data}{long format flowering data}
}
\description{
Get the genotype positions in blocks, trays, and within-tray of the two field experiments
}
\details{
Reads the tsv file of the sown genotypes and replicate id's. Includes quality control from the errors
spotted in situ during the sowing day.
}

