% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_flowering.R
\name{make_data_flowering}
\alias{make_data_flowering}
\title{Read and clean flowering time from the experiment's raw data}
\usage{
make_data_flowering(location)
}
\arguments{
\item{location}{"madrid" or "tuebingen"}
}
\value{
Data frame of flowering times of a specific location.
}
\description{
Returns a long format data frame with information on genotypes, replicates, blocks and
flowering data with its associated quality information (see details).
}
\details{
The function generates 3 additional columns that are added to the genotype
positions \code{\link{genotypepositions}}

FT.q indicates the quality of the pot. Particularly every pot that does not have a plant, or other
symbol of death, would be marked as 0. When it was not clear, for example, when during monitoring
we would see a pot with a blue pin but not flowering date, we would either estimate the date or
give the date when we found this error. When this was particularly unclear, a -9 is assigned
(from comments in the google spreadsheet). Finally, to all those pots that could be assigned
 a flowering date, the number is 1.

FT.date is precissely that date

FT.dif is the value from the date of sowing in the field and the flowering date. Perhaps more
informative could be the date difference between germination time and flowering time, but
germination time has to be assigned a posteriori from image analysis.

NOTE: When merging flowering data information, I kept flag all.x=T for the genotype locations.
This genotype/replicate positions list contain only positions that were correctly sown
so no data of floweirng time at those
}
\seealso{
\code{\link{cleanflowering}} for the possible flags of dead plants
and \code{\link{genotypepositions}} for the original genotype/replicate/block list.
}

