% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_harvest.R
\name{make_data_harvest}
\alias{make_data_harvest}
\title{Read and clean data of harvested plants from the image analysis outputs}
\usage{
make_data_harvest(location = "madrid")
}
\arguments{
\item{location}{"madrid" or "tuebingen"}
}
\value{
Data frame of harvesting info of a specific location.
}
\description{
Returns a long format data frame with information on genotypes, replicates, blocks and
flowering data with its associated quality information (see details).
}
\details{

}
\seealso{
\url{github.com/MoisesExpositoAlonso/hitfruit} and \url{github.com/MoisesExpositoAlonso/hippo} for image processing analyses
}

