% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_vegetative.R
\name{make_data_vegetative}
\alias{make_data_vegetative}
\title{Read and clean data of vegetative plant information from image analysis outputs}
\usage{
make_data_vegetative(location = "madrid")
}
\arguments{
\item{location}{"madrid" or "tuebingen"}
}
\value{
Data frame of green rosette area of a specific location.
}
\description{
Returns a long format data frame with information on genotypes, replicates, blocks and
flowering data with its associated quality information (see details).
}
\details{

}
\seealso{
\url{github.com/MoisesExpositoAlonso/hitfruit} and \url{github.com/MoisesExpositoAlonso/hippo} for image processing analyses
}

