% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarapa.R
\docType{data}
\name{rapa}
\alias{rapa}
\title{Data of growth traits by Francois Vasseur}
\format{A data frame with  ~30,000 rows and ~25 variables:
\describe{
  \item{accessionid}{the 1001 Genomes IDs of the used accessions}
  \item{T_germ}{common name of the accession}
  \item{T_inf}{of origin}
  \item{T_repro}{in degrees}
  \item{PostInf_duration}{in degrees}
  \item{nSiliques}{genetic grups from ADMIXTURE analysis}
  \item{RosetteArea}{8x5 cells quickpot tray }
  \item{RosetteDM}{position coordinate within the 8x5 cells qp tray}
  \item{InfloDM}{experiment location, Madrid or Tuebingen}
  \item{RootDM}{replicate block within experiment}
  \item{ShootAlloc}{the block -- depends on replicate and treatments}
  \item{RootAlloc}{type of replicate. Individual or an experimental population with maximum 30 individuals}
  \item{ReproAlloc}{watering treatment, high or low}

  \item{LUEinf}{quality of flowering time observation. 0=dead, 1=survived, -9=potential error}
  \item{FvFm_mean}{date assigned by experimenter in the flowering time monitoring}
  \item{FvFm_sd}{flowering time in days from the day of sowing till FT.date}
  \item{FvFm_median}{flowering time in days from the day of sowing till FT.date}
  \item{FvFm_S}{flowering time in days from the day of sowing till FT.date}
  \item{FvFm_Wmax}{flowering time in days from the day of sowing till FT.date}

  \item{RAinf}{where is the image from which the information in the following collumns was extracted}
  \item{ERinf}{date where the image was labeled by a human visually inspecting the information in the image}
  \item{RERinf}{plant organ, either rosette or inflorescence}
  \item{DMinf}{replicate type, should be identical to indpop column}
  \item{GRinf}{}
  \item{RGRinf}{quality of harvested plant. 0= no information existed, 1= was labeled and picture exists or written in the diary}

}}
\source{
see \url{1001genomes.org/} for accession list, see \url{} for the publication of growth traits.
}
\usage{
rapa
}
\description{
A dataset containing different life history phenotypes and fitness estimates
of 517 accessions
}
\keyword{datasets}

