% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flowering.R
\name{read_flowering}
\alias{read_flowering}
\title{Read the flowering time raw data}
\usage{
read_flowering(location = "tuebingen", path = "data-raw")
}
\arguments{
\item{location}{madrid or tuebingen}
}
\description{
Read the flowering time raw data
}
\details{
reads/returns raw data of flowering time dates in the spatial position of the field experiment.
}

