% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.R
\name{relative}
\alias{relative}
\title{Make relative fitness fro Genome Wide Hetterogeneous Selection GWHS}
\usage{
relative(f)
}
\arguments{
\item{f}{fitness vector}

\item{type}{what type of relative fitness is wanted}

\item{log}{logical}
}
\description{
Make relative fitness fro Genome Wide Hetterogeneous Selection GWHS
}
\details{

}

